Rework spacing component
