--- ui/mat/doc.go
@@ -1,9 +1,0 @@
-// Package mat provides a collection of Components that can be used
-// to construct a higher-order user interface.
-//
-// Some of the components in this package are inspired by the
-// Material UI design in terms of look and feel and types, however,
-// not all and not fully, hence this is not a compliant Material UI
-// implementation in any way and will likely diverge further down the
-// road.
-package mat
